==> methyl/clock/03_age_plot.ipynb <==
library(ggplot2)
library(reshape)
library(RColorBrewer)
library(stringr)
library(dplyr)
library(reshape2)
library(grid)

library(doParallel)

registerDoParallel(cores=24)

input_name <- paste0('01_935clock')
output_name <- paste0('predict_01_935clock')


file <- paste0('/disk0/sm/methyl/clock/candyCLOCK/DNAmAge_predict_', input_name, '.txt')
age_table <- read.table(file, sep='\t', header=TRUE)

sample <- read.csv('/disk0/sm/methyl/clock/data/sample_sheet.csv', header=TRUE, skip=6) %>% select(Sample_Name, Age, Gender) %>% rename(gender = Gender, age = Age)


algorithm <- c('Horvath', 'Hannum', 'Levine', 'skinHorvath', 'BLUP', 'EN', 'SM')
go <- age_table

all_diff <- cal_diff(go, algorithm, sample)


p <- make_plot(all_diff, output_name, c('id', 'age', 'gender'))
png(filename=paste0('/disk0/sm/methyl/clock/candyCLOCK/age_', output_name, '.png'),width=1200,height=600,unit="px",bg="transparent")
  grid.arrange(p, ncol=1, nrow=1)
dev.off()

mean(df_diff$SM)
sd(df_diff$SM)
mean(df_diff$SM^2)


age_final <- read.table("/disk0/sm/methyl/2024/03_pollution/total/data/age_algorithm.txt", sep="\t", fill = TRUE, header=TRUE)

# pheno <- read.csv('/disk0/sm/methyl/TEST/1031/test/test4_table.txt', sep='\t', header=T)
# target <- pheno %>% filter(Phenotype == 'Normal')
# age_final <- age_final %>% filter(id %in% target$ID)

m_age_final = melt(age_final)
head(m_age_final, 2)
m_age_final <- m_age_final %>% filter(variable != 'TL')
# summary(age_final)

colr <- c('#FB8072', '#FDB462', '#FFED6F', '#B3DE69', '#80B1D3', '#BC80BD', '#f363db', 'grey', 'black', 'white', 'grey30') # 

p <- ggplot(m_age_final, aes(x=variable, y=value, fill=variable)) + 
  # coord_cartesian(ylim = c(-20, 20)) +
  geom_hline(yintercept=0, color='red', linetype = 2) +
  geom_boxplot(outlier.shape = NA) + #
  labs(
        x = 'age predict method', 
        y = 'age difference') + 
  # scale_fill_manual(values=colr) +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=18),
    axis.text.y = element_text(size=15),
    # plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_text(size=18),
    axis.title.y = element_text(size=18),
    axis.ticks.x = element_blank(), 
    legend.text = element_blank(), 
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
    legend.title = element_blank(),
    legend.position = 'none',
  ) + 
  scale_fill_manual(values=colr) 
  # geom_jitter(shape=1, position=position_jitter(0.3), size=0.1) +
  #  geom_boxplot(width=0.1, outlier.shape = NA) 
  


  ggsave(p, file='/disk0/sm/methyl/clock/mydata/age_final.png', width=14000, height=9000, units='px', dpi=1000)
# dev.off()

a <- mata %>% filter(Phenotype == 'Normal')
summary(a$value)
b <- mata %>% filter(Phenotype == 'Patient')
summary(b$value)

head(data)

data <- merge(age_final, pheno, by.x='id', by.y='ID', all=T)
# head(data)
# data$group <- 'under40'
# data[data$Age >= 40, 'group'] = 'over40'
# mata <- melt(data, id.vars=c('id', 'group', 'Age'), measure.vars=c('SM'))
mata <- melt(data, id.vars=c('id', 'Gender'), measure.vars=c('Horvath'))
mata$Gender <- factor(mata$Gender, levels=c('f', 'm'), labels=c('Female', 'male'))
# mata$group <- factor(mata$group, levels=c('under40', 'over40'))
head(mata, 2)


colr <- c('#FB8072', '#80B1D3') # FFED6F FFED6 80B1D3

p <- ggplot(mata, aes(x=Gender, y=value, fill=Gender)) + 
  coord_cartesian(ylim = c(-50, 50)) +
  geom_hline(yintercept=0, color='red', linetype = 2) +
  geom_boxplot(outlier.shape = NA) + #
  labs(
        x = 'age predict method', 
        y = 'age difference') + 
  scale_fill_manual(values=colr) +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=18),
    axis.text.y = element_text(size=15),
    # plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_text(size=18),
    axis.title.y = element_text(size=18),
    axis.ticks.x = element_blank(), 
    legend.text = element_blank(), 
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
    legend.title = element_blank(),
    legend.position = 'none',
  ) 
 
  # geom_jitter(shape=1, position=position_jitter(0.3), size=0.1) +
  #  geom_boxplot(width=0.1, outlier.shape = NA) 
  


  ggsave(p, file='/disk0/sm/methyl/TEST/1031/result/try4_812/test4/age_final_T.png', width=9000, height=9000, units='px', dpi=1000)
# dev.off()

age <- read.table("/disk0/sm/methyl/TEST/1125_clock/grimAGE/age_table.txt", sep="\t", fill = TRUE, header=TRUE)

names(age)[11:19]

algorithm <- names(age)[11:19]
for (i in algorithm){
  tmp <- age[c('Age', i)]
  names(tmp) <- c('Age', 'predict')
  print(paste(i, " : ", cor(tmp$Age, tmp$predict)))
}

algorithm <- names(age)[11:19]
for (i in algorithm){
  tmp <- age[c('Age', i)]
  names(tmp) <- c('Age', 'predict')
  grob3 <- grobTree(textGrob(paste("Coefficient : ", cor(tmp$Age, tmp$predict)), x = 0.03, y = 0.97, hjust = 0, gp = gpar(col = "#ff4848", fontsize = 17, fontface = "bold")))
  p <- ggplot(data=tmp, aes(x=Age, y=predict)) +
  coord_cartesian(xlim = c(0, 100), ylim = c(0, 100))  +
  scale_x_continuous(breaks = seq(0, 100, 10)) +
  scale_y_continuous(breaks = seq(0, 100, 10)) +
  geom_point(size=0.5) +
    labs(title = i, 
          x = 'actual age', 
          y = 'predicted age') + 
    theme_bw() +
    theme(
    #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
      axis.text.x = element_text(size=15),
      axis.text.y = element_text(size=15),
      plot.title = element_text(hjust = 0.5,size=20,face='bold'),
      axis.title.x = element_text(size=18),
      axis.title.y = element_text(size=18),
      axis.ticks.x = element_blank(), 
      legend.text = element_text(size=10),
      
      panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
      panel.grid.major.y = element_line(linetype = "dotted", colour = "gray70"), 
      panel.grid.minor.x = element_blank(), 
      panel.grid.minor.y = element_blank(), 
      legend.title = element_blank()
      ) +
      geom_smooth(method=lm, se=F, color='#ff0055') +
      annotation_custom(grob3)
  #   scale_colour_manual(values=c('#ff4040')) 
  # ggsave(p, file='/disk0/sm/chip/plot/final/total.png', width=2833, height=1890, units='px')
  ggsave(p, file=paste0('/disk0/sm/methyl/TEST/1023/age_', i, '.png'), width=8000, height=8000, units='px', dpi=1000)
}

####################

library(minfi)
library(dplyr)
library(RColorBrewer)
library(ggplot2)
library(viridis)
library(reshape2)
library(data.table)

!Sample_characteristics_ch1     "age: 34" 
!Sample_geo_accession   "GSM100

raw_col <- '#00224D'
data_col <- '#E74646'
old_col <- '#66C2A5'

# data <- read.table("/home/data/two-way_normalized_beta_value_all_CpGs.txt", sep="\t", fill=TRUE, header=TRUE, row.names=1)
data <- fread("/home/data/test/test_beta.txt", sep="\t", fill=TRUE, header=TRUE)[, 2:721]
head(data)

png(filename="/home/data/test/two-way_densityplot.png", width=800, height=500, unit="px")
densityPlot(as.matrix(data), sampGroups = NULL, main = "", xlab = "Beta", pal = data_col, add = TRUE, legend = TRUE)
dev.off()

##########################

table <- read.table('/disk0/sm/methyl/clock/mydata/S678_annotation.txt', sep='\t', header=TRUE, fill=TRUE)

a <- data.frame(table(table$Relation_to_Island)) %>% filter(Var1 != '')
a$Var1 <- factor(a$Var1 , level=c('Island', 'Shore', 'Shelf', 'OpenSea'))
head(a)

p <- ggplot(a, aes(x=Var1, y=Freq, fill=Var1))+
  geom_bar(stat = "identity", size=.2, width=0.6) +
#   coord_cartesian(ylim = c(0, 450000)) +
  # scale_y_continuous(breaks = seq(0, 1, 0.1), labels = paste0(seq(0, 100, 10), '%')) +
  labs(title = 'Annotation relative to CpG Islands') +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=17),
    axis.text.y = element_text(size=15),
    plot.title = element_text(hjust = 0.5,size=20,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(), 
    legend.text = element_text(size=15),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey90"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(),
    legend.title = element_blank(),
    legend.position = 'none'
        ) +
    scale_fill_manual(values=c('#3754f5', '#79b3ff', '#fbb472', '#fbe272')) # '#4723e6', 
  
ggsave(p, file='/disk0/sm/methyl/clock/mydata/island.png', width=8000, height=6000, units='px', dpi=1000)

b <- data.frame(table(table$UCSC)) %>% filter(Var1 != '')
b$Var1 <- factor(b$Var1 , level=c('TSS1500', 'TSS200', '5UTR', 'exon', '3UTR'))
head(b)

p <- ggplot(b, aes(x=Var1, y=Freq, fill=Var1))+
  geom_bar(stat = "identity", size=.2, width=0.6) +
#   coord_cartesian(ylim = c(0, 450000)) +
  # scale_y_continuous(breaks = seq(0, 1, 0.1), labels = paste0(seq(0, 100, 10), '%')) +
  labs(title = 'Functional annotation') +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=17),
    axis.text.y = element_text(size=15),
    plot.title = element_text(hjust = 0.5,size=20,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(), 
    legend.text = element_text(size=15),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey90"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(),
    legend.title = element_blank(),
    legend.position = 'none'
        ) +
    scale_fill_manual(values=c('#19c736', '#72db48', '#f2ff79', '#ffd857', '#fbc072'))
  
ggsave(p, file='/disk0/sm/methyl/clock/mydata/region.png', width=8000, height=6000, units='px', dpi=1000)
